+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Unreleased
 ==========
 
 - Add undocumented `LOOK_REACTORS` season 5 constant
+- Add `HasNativeId` and `Transferable` traits to `Reactor`
+- Add `Transferable` and `Withdrawable` traits to season 1 and 2 object types
 
 0.12.0 (2023-06-07)
 ===================